clean: refactoring. not tested.
