+++ disc/on_ready/on_ready.py
@@ -1,0 +1,8 @@
+import sys
+sys.path.append('../')
+import disc.bot as bot
+
+@bot.bot.event
+async def on_ready():
+    bot.log("Ready!")
+
