+++ disc/randm_teaming/alphabet_index.py
