+++ disc/randm_teaming/funcs.py
@@ -1,6 +1,9 @@
+import sys
+sys.path.append('../')
+
 import discord
 import random
-import alphabet_index
+import disc.randm_teaming.alphabet_index as alphabet_index
 
 async def move_random(channel_move_from: discord.VoiceChannel, channels_move_to: list[discord.VoiceChannel]):
     # 第一引数のVoiceChannelに参加しているメンバーをランダムでmove_toのボイスチャンネルに均等に分配します。
@@ -61,6 +64,3 @@ async def create_temp_channels(ctx, name: str, num: int, cat: discord.CategoryCh
         ch = await create_vc_in_category(ctx, title, cat)
         channels.append(ch)
     return channels
-
-
-
