+++ disc/vc_notice/funcs.py
@@ -1,18 +1,15 @@
-import discord
-import random_teaming
+import sys
+sys.path.append('../')
 
-def count_people(vc: discord.VoiceChannel):
-    # Return how many people is in voice channel.
-    # If voice channel is None, return -1.
-    if vc == None: return -1
-    return len(vc.voice_states.keys())
+import discord
+import disc.libs.voice_channel as lib
 
 def detect_operation(bf:discord.VoiceState, af:discord.VoiceState):
     # When voice state is updated, then return next operation.
     # Return "start", "end" or "many" normally.
     # Return -1 when error happen.
-    bf_cnt = count_people(bf.channel)
-    af_cnt = count_people(af.channel)
+    bf_cnt = lib.count_people(bf.channel)
+    af_cnt = lib.count_people(af.channel)
 
     is_dif_ch = True
     if bf_cnt != -1 and af_cnt != -1:
@@ -28,17 +25,3 @@ def detect_operation(bf:discord.VoiceState, af:discord.VoiceState):
     if bf_cnt == 0:
         return "end"
     return -1
-
-def detect_ch_id(list, chid: int):
-    tmp = -1
-    default = -1
-    for row in list:
-        if row[0] == "default":
-            default = row[1]
-        if row[0] == str(chid):
-            tmp = row[1]
-
-    if tmp != -1:
-        return tmp
-    else:
-        return default
