+++ main.py
@@ -1,4 +1,1 @@
-import env
-import discord_funcs
-import ping_pong
-import vc_funcs
+import disc.bot
